added err msgs

+++ wk4/jpg2png.sh
@@ -12,6 +12,11 @@ do
     )"
     jpg_name=${old_name}.jpg
     png_name=${old_name}.png
+    if test -f "$png_name"
+    then
+        echo "$png_name already exists" 1>&2
+        exit 1
+    fi
     convert "$jpg_name" "$png_name"
     rm "$jpg_name"
     fi
